Renamed some files so not everything was upgrades.go. That name doesn't really work for everything with our upgrade structure.
